refactoring emplacements des fichiers + reagencement entités
