+++ src/main/java/com/materio/materio_backend/business/service/EquipmentService.java
@@ -1,7 +1,7 @@
-package com.materio.materio_backend.service;
+package com.materio.materio_backend.business.service;
 
-import com.materio.materio_backend.model.BO.EquipmentBO;
-import com.materio.materio_backend.model.entity.Equipment;
+import com.materio.materio_backend.business.BO.EquipmentBO;
+import com.materio.materio_backend.jpa.entity.Equipment;
 
 import java.util.List;
