+++ src/main/java/com/materio/materio_backend/business/service/RoomService.java
@@ -1,7 +1,7 @@
-package com.materio.materio_backend.service;
+package com.materio.materio_backend.business.service;
 
-import com.materio.materio_backend.model.BO.RoomBO;
-import com.materio.materio_backend.model.entity.Room;
+import com.materio.materio_backend.business.BO.RoomBO;
+import com.materio.materio_backend.jpa.entity.Room;
 
 import java.util.List;
