+++ src/main/java/com/materio/materio_backend/jpa/repository/EquipmentRepository.java
@@ -1,6 +1,6 @@
-package com.materio.materio_backend.repository;
+package com.materio.materio_backend.jpa.repository;
 
-import com.materio.materio_backend.model.entity.Equipment;
+import com.materio.materio_backend.jpa.entity.Equipment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EquipmentRepository extends JpaRepository<Equipment, Long> {
